refactor manifest implementation with recursive handlers
