+++ core/plugins/render/hooks/hmr.update.ts
@@ -7,7 +7,7 @@ import { Handler, handlerFor } from "@radish/effect-system";
 import { assertObjectMatch } from "@std/assert";
 import { extname } from "@std/path";
 import { updateManifest } from "../../manifest/manifest.ts";
-import { manifestShape } from "./manifest.update.ts";
+import { manifestShape } from "./manifest/mod.ts";
 
 /**
  * @hooks
