+++ core/plugins/render/hooks/manifest/manifest.update.ts
@@ -1,6 +1,6 @@
 import { io } from "$effects/io.ts";
-import { manifest, manifestPath } from "$effects/manifest.ts";
-import type { ElementManifest, Manifest } from "$effects/render.ts";
+import { manifest } from "$effects/manifest.ts";
+import type { ElementManifest } from "$effects/render.ts";
 import {
   elementsFolder,
   generatedFolder,
@@ -12,15 +12,8 @@ import { fragments, shadowRoot } from "@radish/htmlcrunch";
 import { assert } from "@std/assert";
 import { basename, dirname, extname, relative } from "@std/path";
 import { toPascalCase } from "@std/text";
-import { filename, isParent } from "../../../utils/path.ts";
-import { dependencies } from "../utils/walk.ts";
-
-export const manifestShape = {
-  elements: {},
-  imports: {},
-  layouts: {},
-  routes: {},
-} satisfies Manifest;
+import { filename, isParent } from "../../../../utils/path.ts";
+import { dependencies } from "../../utils/walk.ts";
 
 /**
  * The `manifest/update` hook used by the render plugin to extend the manifest with data about elements, routes, layouts etc.
@@ -189,35 +182,3 @@ export const handleManifestUpdateRenderHook = handlerFor(
     return Handler.continue(entry);
   },
 );
-
-/**
- * @hooks
- * - `io/write` Inserts parser imports in the generated manifest module
- * - `manifest/update`
- *
- * @performs
- * - `io/read`
- */
-export const handleManifest = [
-  /**
-   * Decorator for the io/write handler
-   *
-   * Adds the required parser imports to the generated `manifest.ts` module
-   */
-  handlerFor(io.write, (path, content) => {
-    if (path !== manifestPath) return Handler.continue(path, content);
-
-    content =
-      `import { fragments, shadowRoot } from "@radish/core/parser";\n\n${content}`;
-
-    return Handler.continue(path, content);
-  }),
-  handlerFor(manifest.update, (entry) => {
-    // Return early
-    const returnEarly = /\.(spec|test)\.ts$/;
-    if (returnEarly.test(entry.name)) return;
-
-    return Handler.continue(entry);
-  }),
-  handleManifestUpdateRenderHook,
-];
