+++ core/plugins/render/routes/layouts-and-head.ts
@@ -7,7 +7,7 @@ import { manifest } from "$effects/manifest.ts";
 import { type LayoutManifest, type Manifest, render } from "$effects/render.ts";
 import { dev } from "$lib/environment.ts";
 import { isParent } from "$lib/utils/path.ts";
-import { manifestShape } from "../hooks/manifest.update.ts";
+import { manifestShape } from "../hooks/manifest/mod.ts";
 
 /**
  * @performs
