+++ core/plugins/render/transforms/insert-template/insert-template.test.ts
@@ -5,7 +5,7 @@ import { fragments } from "@radish/htmlcrunch";
 import { assertEquals } from "@std/assert";
 import { join } from "@std/path";
 import { describe, test } from "@std/testing/bdd";
-import { manifestShape } from "../../hooks/manifest.update.ts";
+import { manifestShape } from "../../hooks/manifest/mod.ts";
 import { handleRenderTransformTerminal } from "../../transforms/mod.ts";
 import { handleRenderComponents } from "../../components/component.ts";
 import { handleRenderTransformInsertTemplate } from "./insert-templates.ts";
