+++ core/plugins/render/transforms/insert-template/insert-templates.ts
@@ -3,7 +3,7 @@ import { isElementNode } from "@radish/htmlcrunch";
 import { Handler, handlerFor } from "@radish/effect-system";
 import { manifest } from "$effects/manifest.ts";
 import { type Manifest, render } from "$effects/render.ts";
-import { manifestShape } from "../../hooks/manifest.update.ts";
+import { manifestShape } from "../../hooks/manifest/mod.ts";
 import { transformNode } from "../transform-node.ts";
 
 /**
