+++ core/plugins/render/transforms/transform-node.ts
@@ -2,7 +2,7 @@ import { manifest } from "$effects/manifest.ts";
 import { type Manifest, render } from "$effects/render.ts";
 import { isElementNode, type MNode } from "@radish/htmlcrunch";
 import { assertObjectMatch } from "@std/assert";
-import { manifestShape } from "../hooks/manifest.update.ts";
+import { manifestShape } from "../hooks/manifest/mod.ts";
 import { mountHandlerRegistry } from "../state.ts";
 
 /**
